monitoring-service/requirements.txt
# OpenStack (official SDK)
openstacksdk>=3.0.0

# Prometheus metrics
prometheus-client>=0.19.0

# Database (align with VDI schema)
psycopg2-binary>=2.9.9

# HTTP server for /metrics and health
uvicorn[standard]>=0.27.0
fastapi>=0.109.0

# Config
python-dotenv>=1.0.0
